# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CertificateResponse": {
        "description": "One certificate, without any key material.",
        "properties": {
          "autoRenew": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "daysRemaining": {
            "format": "int64",
            "type": "integer"
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "fingerprintSha256": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "inUse": {
            "type": "boolean"
          },
          "issuer": {
            "type": "string"
          },
          "lastError": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "notBefore": {
            "format": "date-time",
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "renewBeforeDays": {
            "format": "int32",
            "type": "integer"
          },
          "revokedAt": {
            "format": "date-time",
            "type": "string"
          },
          "serialNumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "autoRenew",
          "createdAt",
          "createdBy",
          "domains",
          "id",
          "inUse",
          "name",
          "provider",
          "renewBeforeDays",
          "status",
          "version"
        ],
        "type": "object"
      },
...
      "CreateDomainGroupRequest": {
        "description": "A new domain group.",
        "properties": {
          "description": {
            "maxLength": 512,
            "minLength": 0,
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "parentId": {
            "format": "uuid",
            "type": "string"
          },
          "slug": {
            "minLength": 1,
            "pattern": "^[a-z0-9]([a-z0-9-]{0,62}[a-z0-9])?$",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug"
        ],
        "type": "object"
      },
...
      "DomainGroupResponse": {
        "description": "One domain group.",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "depth": {
            "format": "int32",
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "memberCount": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "format": "uuid",
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          },
          "yourLevel": {
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "createdBy",
          "depth",
          "id",
          "memberCount",
          "name",
          "path",
          "version"
        ],
        "type": "object"
      },
...
      "NginxInstanceResponse": {
        "description": "A managed NGINX host.",
        "properties": {
          "agentBaseUrl": {
            "type": "string"
          },
          "agentCertFingerprint": {
            "type": "string"
          },
          "agentVersion": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "lastSeenAt": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nginxVersion": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "agentBaseUrl",
          "agentCertFingerprint",
          "createdAt",
          "environment",
          "hostname",
          "id",
          "name",
          "status",
          "version"
        ],
        "type": "object"
      },
...
      "RegisterNginxInstanceRequest": {
        "properties": {
          "agentBaseUrl": {
            "minLength": 1,
            "type": "string"
          },
          "agentCertFingerprint": {
            "minLength": 1,
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "hostname": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "agentBaseUrl",
          "agentCertFingerprint",
          "hostname",
          "name"
        ],
        "type": "object"
      },
...
      "RequestAcmeRequest": {
        "properties": {
          "autoRenew": {
            "type": "boolean"
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "type": "array"
          },
          "name": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "renewBeforeDays": {
            "format": "int32",
            "maximum": 89,
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "domains",
          "name"
        ],
        "type": "object"
      },
...
      "RotateAgentCertificateRequest": {
        "properties": {
          "agentCertFingerprint": {
            "maxLength": 95,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "agentCertFingerprint"
        ],
        "type": "object"
      },
...
      "UpdateDomainGroupRequest": {
        "description": "Changes to a domain group.",
        "properties": {
          "description": {
            "maxLength": 512,
            "minLength": 0,
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer-jwt": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Management API for reverse-proxy sites across one or more NGINX hosts.\n\nAuthentication is OIDC: obtain an access token from Keycloak and send it as\na bearer token. Errors are RFC 9457 problem documents; branch on the stable\n`type` URI rather than on the human-readable `detail`.\n",
    "license": {
      "name": "Proprietary"
    },
    "title": "Easy NGINX Admin API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/certificates": {
      "get": {
        "description": "Metadata only. Private key material is never returned by any endpoint.",
        "operationId": "list_4",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CertificateResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List certificates",
        "tags": [
          "Certificates"
        ]
      },
      "post": {
        "description": "Validates with HTTP-01 against the managed NGINX hosts. Wildcards need DNS-01 and are refused. Requires MANAGE over the namespace of every domain requested, because a certificate is authority to terminate TLS for that name.",
        "operationId": "request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAcmeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Request a certificate over ACME",
        "tags": [
          "Certificates"
        ]
      }
    },
...
    "/api/v1/domain-groups": {
      "get": {
        "operationId": "list_3",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DomainGroupResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List domain groups visible to you",
        "tags": [
          "Domain groups"
        ]
      },
      "post": {
        "description": "A top-level group requires global MANAGE; a subgroup requires MANAGE on its parent.",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDomainGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DomainGroupResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Create a domain group",
        "tags": [
          "Domain groups"
        ]
      }
    },
    "/api/v1/domain-groups/{id}": {
      "delete": {
        "description": "Refused while the group has subgroups or members, because deleting it would silently revoke every grant made over it.",
        "operationId": "delete_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Delete an empty domain group",
        "tags": [
          "Domain groups"
        ]
      },
      "get": {
        "operationId": "get_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DomainGroupResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Fetch one domain group",
        "tags": [
          "Domain groups"
        ]
      },
      "put": {
        "description": "The path is identity and cannot be changed: moving a group would silently change which sites every grant beneath it reaches.",
        "operationId": "update_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDomainGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DomainGroupResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Rename a domain group",
        "tags": [
          "Domain groups"
        ]
      }
    },
    "/api/v1/domain-groups/{id}/members": {
      "get": {
        "operationId": "members",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List the sites in a group",
        "tags": [
          "Domain groups"
        ]
      }
    },
    "/api/v1/domain-groups/{id}/members/{siteId}": {
      "delete": {
        "operationId": "removeMember",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "siteId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Remove a site from a group",
        "tags": [
          "Domain groups"
        ]
      },
      "post": {
        "description": "Requires MANAGE on both the group and the site. Requiring only the group would let someone add a site they have no rights to and inherit control of it.",
        "operationId": "addMember",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "siteId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Add a site to a group",
        "tags": [
          "Domain groups"
        ]
      }
    },
...
    "/api/v1/nginx-instances": {
      "get": {
        "operationId": "list_2",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NginxInstanceResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List managed NGINX instances",
        "tags": [
          "NGINX instances"
        ]
      },
      "post": {
        "description": "Restricted to SUPER_ADMIN: registering a host decides where configuration is deployed.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterNginxInstanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NginxInstanceResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Register an NGINX instance",
        "tags": [
          "NGINX instances"
        ]
      }
    },
    "/api/v1/nginx-instances/{id}": {
      "get": {
        "operationId": "get_2",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NginxInstanceResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Fetch one NGINX instance",
        "tags": [
          "NGINX instances"
        ]
      }
    },
    "/api/v1/nginx-instances/{id}/agent-certificate": {
      "put": {
        "description": "The fingerprint is pinned in addition to CA verification, so replacing an agent's certificate makes the host unreachable until this is called. Swap the certificate on the host first, then call this: the window between the two is a loss of control, never of traffic \u2014 the host keeps serving throughout.",
        "operationId": "rotateAgentCertificate",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RotateAgentCertificateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NginxInstanceResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Trust a new agent certificate for this host",
        "tags": [
          "NGINX instances"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer-jwt": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
...
}